+++ backend/app/Http/Controllers/BudgetController.php
@@ -6,10 +6,14 @@ use App\Http\Requests\BudgetRequest;
 use App\Http\Resources\BudgetResource;
 use App\Models\Budget;
 use App\Services\BudgetService;
+use App\Models\Account;
+// use App\Services\BudgetService;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class BudgetController extends Controller
 {
+    use AuthorizesRequests;
     use ApiResponse;
 
     public function __construct(
@@ -18,6 +22,7 @@ class BudgetController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Budget::class);
         return $this->success(
             BudgetResource::collection(
                 $this->budgetService->index()
@@ -27,6 +32,7 @@ class BudgetController extends Controller
 
     public function store(BudgetRequest $request)
     {
+        $this->authorize('create', Budget::class);
         return $this->success(
             new BudgetResource(
                 $this->budgetService->store(
@@ -40,6 +46,7 @@ class BudgetController extends Controller
 
     public function show(Budget $budget)
     {
+        $this->authorize('view', $budget);
         return $this->success(
             new BudgetResource(
                 $this->budgetService->show($budget)
@@ -49,6 +56,7 @@ class BudgetController extends Controller
 
     public function destroy(Budget $budget)
     {
+        $this->authorize('delete', $budget);
         $this->budgetService->destroy($budget);
 
         return $this->success(
