+++ backend/app/Http/Controllers/TransferController.php
@@ -7,11 +7,12 @@ use App\Http\Resources\TransferResource;
 use App\Models\Transfer;
 use App\Services\TransferService;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
 
 class TransferController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     public function __construct(
         private TransferService $transferService
@@ -19,6 +20,7 @@ class TransferController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Transfer::class);
         return $this->success(
             TransferResource::collection(
                 $this->transferService->index()
@@ -28,6 +30,7 @@ class TransferController extends Controller
 
     public function store(TransferRequest $request)
     {
+        $this->authorize('create', Transfer::class);
         $transfer = $this->transferService->store(
             $request->validated()
         );
@@ -41,6 +44,7 @@ class TransferController extends Controller
 
     public function show(Transfer $transfer)
     {
+        $this->authorize('view', $transfer);
         abort_if(
             $transfer->user_id !== Auth::id(),
             403
@@ -58,6 +62,7 @@ class TransferController extends Controller
 
     public function destroy(Transfer $transfer)
     {
+        $this->authorize('delete', $transfer);
         $this->transferService->delete($transfer);
 
         return $this->success(
